Foreign Keys moved from collections to object types.
Type representations are no longer optional

Schema Response now includes a reference to the scalar type to be used for count results.

AggregateFunctionDefinition is now an enum, so we map based on function name. Note! We are currently lying by omission about the return types. Postgres aggregates will return NULL if aggregating over no rows, except COUNT.

We should have a discussion about wether we want to change aggregate function definitions to reflect this behavior, whether all these scalars will be implicitly nullable, or whether we want to change the SQL using COALESCE to default to some value when no rows are present.

Arguably, there's no proper MAX, MIN, or AVG default values.
As for SUM, ndc-test expects all SUM return values to be either represented as 64 bit integers or 64 bit floats. Postgres has types like INTERVAL, which is represented as a string, and can be aggregated with SUM.

We need to discuss whether any of the above needs to be revisited. We cannot represent intervals as float64 or int64.