Remove mold suggestion in `.envrc.example` (#688)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

It's totally reasonable that a user would just copy `.envrc.example` to
`.envrc` and then get going, however this means we try and use `mold`,
which we do not install in the Nix flake, and does not work on MacOS.
Let's make the default path as conservative and likely to work as
possible.

### How

Remove `mold` advice from `.envrc.example`, make `CARGO_INCREMENTAL`
flag opt in.