Add subquery to separate where clause from lateral joins used for relationships (#713)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->
This PR copies some patterns used in v2, which it turns out yield better
performance on cockroachdb.

<!-- Consider: do we need to add a changelog entry? -->

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

We add an additional subquery wherever we directly reference a source
table, to make sure the where clause for that table is applied before
any lateral joins.

### Ordering

Ordering and limit are also moved into the subquery.

V2 instead does the following:
- if ordering across a relationship, put the join and order by clause
outside the subquery, at the same level as other relationships. also put
any limit clause there
- if ordering on a local field, put the order by and limit inside the
subquery. additionally, also put a copy of the order by clause outside
the subquery, at the same level as relationships

We do not replicate this behavior, and the last bit worries me.
Will ordering be preserved without the second order by clause?

Commiting as is for now but will probably add the clause.

---------

Co-authored-by: Daniel Harvey <danieljamesharvey@gmail.com>