Improve ordering with subquery (#715)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

This is a follow up to #713

That PR adds a subquery to help query planners apply the where clause of
a table before any lateral joins required for relationships.

That work introduced correctness issues:
```sql
SELECT
"%1_Album"."Title" AS "Title"
FROM
  (
    SELECT
      "%0_Album".*
    FROM
      "public"."Album" AS "%0_Album"
    ORDER BY
      "%0_Album"."AlbumId" ASC
    LIMIT
      5
  ) AS "%1_Album"
```
The above looks fine, but because that top level query may be joined to
other tables for relationships, ordering is actually not guaranteed.
We add the order by clause at the top level too:
```sql
SELECT
"%1_Album"."Title" AS "Title"
FROM
  (
    SELECT
      "%0_Album".*
    FROM
      "public"."Album" AS "%0_Album"
    ORDER BY
      "%0_Album"."AlbumId" ASC
    LIMIT
      5
  ) AS "%1_Album"
ORDER BY
  "%1_Album"."AlbumId" ASC
```
We do still want the order by and limit clauses in the subquery so we
reduce the rows in scope as early as possible.

When ordering via related tables, we put the order by and limit clauses
in the parent query:
```sql
SELECT
  "%1_Album"."Title" AS "Name"
FROM
  (
    SELECT
      "%0_Album".*
    FROM
      "public"."Album" AS "%0_Album"
  ) AS "%1_Album"
  LEFT OUTER JOIN LATERAL (
    SELECT
      "%2_ORDER_PART_Artist"."Name" AS "Name"
    FROM
      (
        SELECT
          "%2_ORDER_PART_Artist"."Name" AS "Name"
        FROM
          "public"."Artist" AS "%2_ORDER_PART_Artist"
        WHERE
          (
            "%1_Album"."ArtistId" = "%2_ORDER_PART_Artist"."ArtistId"
          )
      ) AS "%2_ORDER_PART_Artist"
  ) AS "%3_ORDER_FOR_Album" ON ('true')
ORDER BY
  "%3_ORDER_FOR_Album"."Name" ASC
LIMIT
  5 OFFSET 3
```

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->

<!-- Consider: do we need to add a changelog entry? -->

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->
When ordering using columns in other tables, 
When ordering through a relationship, put the order by clause and limit
in the parent query.
When ordering from local column, put the order by clause and limit in
the subquery.
Additionally, add a copy of the order by clause in the parent, to ensure
correctness.

See code comments for details