Remove haskell CI so we can better integrate with rest of CI, which is bash scripts (#734)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->
Remove haskell CI step, so future PRs can better integrate with rest of
CI which is bash scripts.
<!-- Consider: do we need to add a changelog entry? -->

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

`changelog.hs` would do the following to changelog.md:
1. remove the top 4 lines
```md
# Changelog

## [Unreleased]

```
2. Prepend the text with the following changelog template:
```md
# Changelog

## [Unreleased]

### Added

### Changed

### Fixed

## [<new-version>] - <date>
```
3: Find a line that starts with `[Unreleased]: ` and replace it with a
reference to the newly created tag, and update the reference to
unreleased code

We replace this with a bash script that does much of the same. Instead
of deleting the top 4 lines, we replace the `## [Unreleased]` header
with the new template.

This change is a no-op, but will facilitate upcoming changes to CI.