Remove python ci script (#736)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->
We had a python script used during the shipping workflow to generate
release notes from the changelog.

This PR removes that script to better integrate with future ci scripts
that will also be in bash.
<!-- Consider: do we need to add a changelog entry? -->

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->
We create a bash script that takes 3 inputs: the new version tag, the
input file (changelog.md), and the output file.

The script will write directly to the output file, if all goes well.
This is different from the python script, which printed to stdout so the
output could be captured later.

The script itself is rather simple:
- Match changelog lines between the version header and the next header
or end marker
- Error if unable to match any lines
- Trim trailing and leading empty lines
- Print the changelog lines and some boilerplate to the output file.
Note this will overwrite the file if present