Support setting array columns to empty array instead of setting those to null (#762)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->

Attempting to set an array column to `[]` would instead set the column
to `NULL`.
This happened because we process the incoming values using `array_agg`,
which returns null if aggregating over no rows.

<!-- Consider: do we need to add a changelog entry? -->

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->
We wrap the array aggregation in `COALESCE`, with a default value of an
empty array cast to the right type.

The new behavior when updating an array column is now
- setting to an array with elements works as before
- setting to null works as before
- setting to an empty array works as expected, wheras before it would be
set to null.

The PR also adds snapshot tests for all 3 cases. The snapshotted SQL has
been tested and verified to behave as intended.

Some other snapshots are also updated due to the change in generated
SQL.
It seems we use the functions in `values.rs` every time we have some
kind of json input.

This is not what I set out to fix, and there could be issues if we're
relying on the null behavior here, but I don't think we are.