refactor state to prepare for dynamic conection pools (#782)
This PR refactors state in preparation for dynamic connection pools. A
brief list of changes follows:

- update ndc-models dependency to 0.2.4
- update ndc-sdk depndency to 0.7.0
- configuration.rs: update runtime Configuration struct: replace
connection_uri string with a ConnectionSettings enum, which will have a
variant per connection pooling mode. For now it only has static, which
is default and is the same as the existing behavior.
- normalize accessing the environment during startup:
  - previous behavior had us accessing the environment in two places:
- when creating runtime configuration from parsed configuration, we
would potentially read the env to get a connection string to store in
configuration
- when creating a connection pool while creating state, we would
potentially read the env to get ssl information for pool creation
- this becomes an issue for the upcoming feature allowing creating
connection pools after startup.
- therefore, we change the runtime configuration to include all
information we may read from the env. Specifically ssl information is
added there
- as a result we change how pools are created, they now expect the ssl
information instead of the environment (connect.rs)"
- we introduce the Redacted struct. This manually implements Display and
Debug to show `[REDACTED]` for both. This is a safety feature in case
our state gets printed somewhere. In theory this should never be
relevant but no harm implemeting this.
- introduce PoolManager enum and PoolContext struct.
- PoolManager is an abstraction that will handle the various connection
modes. For now it onlys has the Static mode which is identical to
current behavior. It is stored in the State.
- a PoolContext can be aquired from the PoolManager. This contains a
connection pool, and DatabaseInfo for the connection pool. (host, port,
username to be put in traces)
- the pool manager also has abstractions for updating metrics for
(potentially multiple) connection pools
- refactor pool metrics to be labeled
- given we can now have multiple connection pools, pools metrics are now
labeled so we can track metrics for each individual connection pool