Update CI to automate release process (#784)
Automate CI

This PR adds workflows to automate CI.
Once merged, we can trigger the "create release" workflow to create a
release PR. When that is merged, the new version is tagged and released
automatically.

We also inline the cli plugin definitions. To do this, we deprecate
using the postgres cli to generate the metadata for ci, and instead we
use templates in ci.