Upload ci artifact so it can be used in next step (#795)
We had a bug where the reate-ndc-hub-pr: job would failt to find the
connector-definition.tgz artifact, because we never uploaded it.

This PR adds that step.

This PR also prevents envsubst from substituting env vars we don't want
to replace.
This was an issue with some of the scripts that are in the metadata file